Reorder colors and improve logs
