+++ src/python/orbit_determination/mod.rs
@@ -16,20 +16,14 @@
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-use std::collections::HashMap;
-
-use crate::cosmic::Cosm;
 use crate::io::tracking_data::DynamicTrackingArc;
-use crate::io::trajectory_data::TrajectoryLoader;
 use crate::io::ExportCfg;
-use crate::od::msr::RangeDoppler;
 use crate::od::noise::GaussMarkov;
 use crate::od::process::FltResid;
-use crate::od::simulator::TrackingArcSim;
 pub use crate::od::simulator::TrkConfig;
 pub use crate::{io::ConfigError, od::prelude::GroundStation};
-use crate::{NyxError, Spacecraft};
 use pyo3::{prelude::*, py_run};
+mod arc;
 pub(crate) mod estimate;
 mod ground_station;
 mod process;
@@ -42,7 +36,7 @@ pub(crate) fn register_od(py: Python<'_>, parent_module: &PyModule) -> PyResult<
     let sm = PyModule::new(py, "_nyx_space.orbit_determination")?;
 
     sm.add_class::<GroundStation>()?;
-    sm.add_class::<GroundTrackingArcSim>()?;
+    sm.add_class::<arc::GroundTrackingArcSim>()?;
     sm.add_class::<DynamicTrackingArc>()?;
     sm.add_class::<TrkConfig>()?;
     sm.add_class::<OrbitEstimate>()?;
@@ -59,56 +53,4 @@ pub(crate) fn register_od(py: Python<'_>, parent_module: &PyModule) -> PyResult<
     );
     parent_module.add_submodule(sm)?;
     Ok(())
-}
-
-#[derive(Clone)]
-#[pyclass]
-pub struct GroundTrackingArcSim {
-    inner: TrackingArcSim<Spacecraft, RangeDoppler, GroundStation>,
-}
-
-#[pymethods]
-impl GroundTrackingArcSim {
-    /// Initializes a new tracking arc simulation from the provided devices, trajectory, and the random number generator seed.
-    #[new]
-    pub fn with_seed(
-        devices: Vec<GroundStation>,
-        trajectory: TrajectoryLoader,
-        configs: HashMap<String, TrkConfig>,
-        seed: u64,
-    ) -> Result<Self, NyxError> {
-        // Try to convert the dynamic trajectory into an Orbit trajectory
-        let traj = trajectory
-            .to_traj()
-            .map_err(|e| NyxError::CustomError(e.to_string()))?;
-
-        let inner = TrackingArcSim::with_seed(devices, traj, configs, seed)
-            .map_err(NyxError::ConfigError)?;
-
-        Ok(Self { inner })
-    }
-
-    /// Generates simulated measurements and returns the path where the parquet file containing said measurements is stored.
-    /// You may specify a metadata dictionary to be stored in the parquet file and whether the filename should include the timestamp.
-    #[pyo3(text_signature = "(path, metadata=None, timestamp=False)")]
-    pub fn generate_measurements(
-        &mut self,
-        path: String,
-        export_cfg: ExportCfg,
-    ) -> Result<String, NyxError> {
-        let cosm = Cosm::de438();
-        let arc = self.inner.generate_measurements(cosm)?;
-
-        // Save the tracking arc
-        let maybe = arc.to_parquet(path, export_cfg);
-
-        match maybe {
-            Ok(path) => Ok(format!("{}", path.to_str().unwrap())),
-            Err(e) => Err(NyxError::CustomError(e.to_string())),
-        }
-    }
-
-    pub fn __repr__(&self) -> String {
-        format!("{}", self.inner)
-    }
 }
